requirements.txt
django==2.2
django-ckeditor==6.1.0
django-notifications-hq==1.6.0
Pillow==8.3.1
mysqlclient==2.0.3